Make authentication (login and logout) work.
