+++ src/Login/Actions.js
@@ -1,6 +1,4 @@
-import { startAuth } from '../Authentication/Actions';
 import Fetch from '../store/Fetch';
-import { put, call } from 'redux-saga/effects';
 import useSwr from 'swr';
 
 /* UPDATE FIELDS */
@@ -18,12 +16,13 @@ const DATABASE_TYPE = "database";
 
 export const useUser = (id) => {
   const shouldFetch = id >= 0 && id!= null;
-  const { data, error } = useSwr(shouldFetch ? `/api/users/${id}` : null);
+  const { data, mutate, error } = useSwr(shouldFetch ? `/api/users/${id}` : null);
 
   return {
     user: data,
     isLoading: !error && !data,
-    isError: error
+    isError: error,
+    mutate
   }
 }
 
@@ -42,60 +41,3 @@ export const newLogin = (username, password) => {
 
   return data;
 }
-
-export default function login(username, password){
-   return startAuth(DATABASE_TYPE, null, "", username, password);
-}
-
-
-/*
-export function requestRegister(registerType, code, values, redirectURL){
-
-  return {
-    type: REGISTER_START,
-    registerType: registerType,
-    code: code,
-    values: values,
-    redirectURL: redirectURL
-  };
-}
-
-export function* register(action){
-
-  const {
-    code,
-    values,
-    redirectURL,
-    registerType
-  } = action;
-
-  let jsonData = null;
-  let registerData = null;
-
-  if(registerType === "database"){
-    registerData = {
-      username: values.username,
-      password: values.password,
-      email: values.email,
-      name: "",
-      type: registerType
-    };
-  }
-  else {
-    registerData = {
-      code: code,
-       type: registerType,
-       redirectURL: redirectURL
-    };
-  }
-
-  jsonData = JSON.stringify(registerData);
-
-  try {
-    const postData = yield call(Fetch.POST, "/api/users", [], jsonData);
-    yield put({ type: REGISTER_SUCCESS, data: postData });
-  }
-  catch(error){
-    yield put({ type: REGISTER_FAILURE });
-  }
-}*/
