+++ src/Story/Container.js
@@ -1,12 +1,11 @@
 import { connect } from 'react-redux'
-import { REQUEST_STORY, REGISTER_TEMP_USER } from './Actions';
+import { REGISTER_TEMP_USER } from './Actions';
 import Story from './SWRComponent';
 import getOptionsValues from '../utils/misc';
 
 const mapStateToProps = (state) => {
   return {
     loggedIn: state.session.loggedIn,
-    oauthServices: state.config.oauth,
     commentOptions: getOptionsValues(state.config.options, 'comments'),
     userExists: state.session.user.loaded
   }
