+++ src/sagas/index.js
@@ -2,7 +2,6 @@ import { all, call } from 'redux-saga/effects';
 import mainSaga from './mainSaga';
 import storiesSaga from './storiesSaga';
 import categoriesSaga from './categoriesSaga';
-import authSaga from './authSaga';
 import composerSaga from './composerSaga';
 import profileSaga from './profileSaga';
 import commentSaga from './commentSaga';
@@ -12,7 +11,6 @@ import draftsSaga from './draftsSaga';
 export default function* rootSaga(dispatch){
     yield all([
       call(mainSaga, dispatch),
-      call(authSaga),
       call(storiesSaga),
       call(categoriesSaga),
       call(composerSaga),
