+++ src/sagas/mainSaga.js
@@ -2,7 +2,6 @@ import { takeEvery } from 'redux-saga/effects';
 import {
   loadWebsocket,
   REQUEST_INITIALIZE_WS,
-  REQUEST_CONFIG_DATA
 } from '../App/Actions';
 
 export default function* mainSaga(dispatch) {
